refactor(api): extract postManagedComment service from the github-comment route
Moves the handler body of POST /v1/:workspace/github/comment into
apps/api/src/github-comment-service.ts (postManagedComment) verbatim, so
the same gather/check/upsert logic can be called in-process from the
hosted MCP server. The route is now a thin parse + rate-limit + scope
wrapper. Existing route tests pass unchanged.
